deletes and get ones fixed

+++ src/services/autor/DeleteAutorService.ts
@@ -1,6 +1,7 @@
 import { getRepository } from "typeorm";
 import { validate } from "uuid";
 import { Autor } from "../../entities/Autor";
+import { ObraAutor } from "../../entities/ObraAutor";
 
 export class DeleteAutorService {
     async execute(id: string) {
@@ -12,6 +13,13 @@ export class DeleteAutorService {
         
         if(!autor){
             return new Error("Autor não existe!");
+        }
+
+        const repoObraAutor = getRepository(ObraAutor);
+        const obraWithObraAutor = await repoObraAutor.findOne({where: {autor_id : id}});
+
+        if(obraWithObraAutor){
+            return new Error("Autor com obras cadastradas");
         }
 
         await repo.delete(id);
