+++ src/services/disciplina_versao/DeleteDisciplinaVersaoService.ts
@@ -1,5 +1,6 @@
 import { getRepository } from "typeorm";
 import { validate } from "uuid";
+import { Bibliografia } from "../../entities/Bibliografia";
 import { DisciplinaVersao } from "../../entities/DisciplinaVersao";
 
 export class DeleteDisciplinaVersaoService {
@@ -12,6 +13,13 @@ export class DeleteDisciplinaVersaoService {
         
         if(!versao){
             return new Error("Versão de Disciplina não existente!");
+        }
+
+        const repoBibliografia = getRepository(Bibliografia);
+        const versaoWithBibliografias = await repoBibliografia.findOne({where: {versao_disciplina_id : id}});
+
+        if(versaoWithBibliografias){
+            return new Error("Versão de disciplina com bibliografias cadastradas");
         }
 
         await repo.delete(id);
