+++ src/services/obra/DeleteObraService.ts
@@ -1,6 +1,8 @@
 import { getRepository } from "typeorm";
 import { validate } from "uuid";
+import { Bibliografia } from "../../entities/Bibliografia";
 import { Obra } from "../../entities/Obra";
+import { ObraAutor } from "../../entities/ObraAutor";
 
 export class DeleteObraService {
     async execute(id: string) {
@@ -12,6 +14,20 @@ export class DeleteObraService {
         
         if(!obra){
             return new Error("Obra não existe!");
+        }
+
+        const repoBibliografia = getRepository(Bibliografia);
+        const obraWithBibliografias = await repoBibliografia.findOne({where: {obra_id : id}});
+
+        if(obraWithBibliografias){
+            return new Error("Obra com bibliografias cadastradas");
+        }
+
+        const repoObraAutor = getRepository(ObraAutor);
+        const obraWithObraAutor = await repoObraAutor.findOne({where: {obra_id : id}});
+
+        if(obraWithObraAutor){
+            return new Error("Obra com autores cadastrados");
         }
 
         await repo.delete(id);
